Move system lifecycle endpoints to seperate service part
